Move content api things to relevant package
