+++ src/content_api/details_utils.py
